Make the hostname mandatory in the configuration
If the plugin is installed without any configuration it defaults the
hostname to localhost, assuming that Graphite is running there.

In deployments where the plugin is bundled with Gerrit, but Graphite is
not necessarily by default on the same host, this causes the logs to be
spammed with "connection refused" errors.

Make the hostname mandatory, so that if the Graphite server is actually
running at localhost it must be explicitly set in the config.

Change-Id: I3a7651f229626e2180717a4dd105f36c34433bc1